gitrecombo/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
openai>=1.3.0
jinja2>=3.1.2
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
